+++ consts.js
@@ -1,3 +1,3 @@
 export const name = '@shgysk8zer0/components';
-export const version = '0.0.1';
+export const version = '0.0.3';
 export const url = `https://unpkg.com/${name}@${version}/`;